requirements.txt
pandas
numpy
scikit-learn
scipy
streamlit
matplotlib
seaborn
plotly
streamlit-option-menu
streamlit_extras
streamlit-authenticator
deta==1.1.0
xgboost
openpyxl
supabase
python-dotenv